requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.0
bcrypt==4.1.1
python-jose==3.3.0
PyJWT==2.12.1
passlib==1.7.4
email-validator==2.1.0
tenacity==8.2.3
langdetect==1.0.9
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
imbalanced-learn==0.11.0
